packages/sage-data-gen/pyproject.toml
[project]
name = "sage-data-gen"
version = "0.1.0"
description = "Data generation pipelines for sage-benchmark"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "datasets",
    "Pillow",
    "tqdm",
    "rank-bm25>=0.2.2",
    "sage-benchmark",
    "sage-llm",
    "duckdb>=1.5.1",
    "ftfy>=6.0",
]

[project.scripts]
sagegen = "sage_data_gen.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sage_data_gen"]
artifacts = ["sage_data_gen/**/*.jsonl"]

[tool.uv.sources]
sage-benchmark = { workspace = true }
sage-llm = { workspace = true }
